Create the specified prefix directory automatically
Although WINE will create the prefix directory by itself if that
path's parent exists, it will not do so if the prefix is more than
a level deep in uncreated directories.

For example, if the user is creating a new prefix in ~/wine/application
and ~/wine does not exist, WINE will not create the path and any
attempts to start applications with that prefix will fail.